flux-resource: drain: suggest --force when reason exists
Problem: `flux resource drain TARGET REASON` fails with an error
if the TARGET is already drained without updating the reason, but it
may be helpful to the user to suggest a use of --force to overwrite
an existing drain reason.

Check for errno.EEXIST when drain fails, and suggest using --force
to overwrite reason in case this is what the user wanted to do.

Fixes #4922