flux-queue: adjust default output of min/max/range
Problem: The current default outputs of node, core, and gpu values
and ranges in flux queue list is not considered very user friendly.

Solution: If the value of nnodes, ncores, or ngpus is not configured,
instead of always returning "inf", return "inf" only if it is the max.
Return 0 if it is the min.  As a result, the default output from the
range values of nodes, cores, and gpus becomes 0-inf.

Update test in t2241-queue-cmd-list.t accordingly.

Fixes #4927