job-manager: add job.inactive-add, inactive-remove
Problem: jobtap plugins are not informed when jobs are added
or removed from the inactive hash.

The original jobtap design predates the job manager's caching of
inactive jobs, so this is as designed.

Add two new callbacks to support plugins that track inactive jobs:

  job.inactive-add
    Invoked 1) when an inactive job is loaded from the KVS on restart
    and (2) when a job transitions from active to inactive.

  job.inactive-remove
    Invoked when an inactive job is purged.

Note: when a jobtap plugin is tracking all jobs, it will get both a
job.new and a job.inactive-add for new jobs, and only a job.inactive-add
for jobs recovered on restart.

Update t2212-job-manager-plugins.t to
- look for 21 callbacks not 20 in args test
- verify that job.inactive-add is called on inactive transition
- verify that job.inactive-remove is called on purge