modules/sdbus: new service
Problem: integrating reactive sdbus application code with Flux
is painful.

Add a service that can act as a bridge between the flux broker
and the local D-Bus daemon.  One connection to D-Bus is shared
among one or more Flux users.

An sdbus.call RPC maps D-Bus RPCs to flux RPCs:
- a Flux request is translated to a D-Bus method-call
- a D-Bus method-return or method-error message is translated to a flux
  response
- multiple outstanding requests are managed using D-Bus cookies

An sdbus.subscribe RPC lets Flux consume D-Bus signal messages
as a stream of responses to the subscribe request.

JSON is translated to sdbus objects and vice versa.  A generic translator
is not provided at this time.  Simple messages are handled generically,
but complex messages require custom translation functions.  A few custom
translators are provided for some complex messages Flux will need when
dealing with transient units.  The translation follows a pattern, and is
not "improved" to make prettier or more self-describing JSON so that the
door to a full generic translator is open for later.