libsubprocess/server: use a list instead of a hash
Problem: the subprocess server tracks subprocesses in a hash
by pid, but before the subprocess is assigned a pid, it is
not tracked by the server.

In theory, teardown of the subprocess server at the wrong moment
could lead to leaked subprocesses.

Use a list instead of a hash, and add the subprocess to the
container immediately upon creation.  The downside of this change
is that rexec.kill and rexec.output must now perform a linear search
to look up the subprocess by pid.

Rename some functions that were pid-hash specific.