libsubproces/server: improve rexec.exec errors
Problem: the rexec.exec RPC handler logs errors on failure
and does not return human readable errors to the requestor.

Change the handler to send textual errors back to the requestor
and avoid logging except in the case of a response failure.

Fix a case where a response failure was handled by sending an error
response, which is counter to established practice.