libsubprocess/server: rename fatal error function
Problem: the internal function "internal_fatal" is for subprocesses
not for the entire subprocess server as might be assumed, given its
name and the fact that its first argument is the server context.

Rename to proc_internal_fatal() and change parameters to take only
a (flux_subprocess_t *).  Hopefully this makes the purpose of the
function more clear to casual readers.