python: add RFC 36 submission directive parser
Problem: RFC 36 describes a specification for embedding submission
options in a script or file, but there is no implementation for
processing a file which may contain these directives.

Add a new Python flux.job.directives module to assist submission
utilities in extracting directives from submission scripts and other
files.  The implementation provides a DirectiveParser class which,
given an input file, returns a possibly empty list of directives and
the ingested script.

The DirectiveParser looks for sentinels matching the pattern specified
in RFC 36 and extracts each line or quoted multiline into a Directive
object. Single line strings are split into multiple tokens by Python's
shlex module, which allows lines to contain familiar shell quoting and
comments.  As a convenience, inline triple quoting is also supported.

Directive objects represent a single action and its arguments. The
only directive types supported in this version are SETARGS, which
include one or more command line arguments in args, and NOOP, which
represents an empty directive. (A NOOP occurs when a directive is
quoted in full). Other directive actions may be added in the future.