doc: add shell help on flux-jobs(1) formatting
Problem: The exclamation point used in flux-jobs(1) conversion flags
is interpreted by some shells, such as csh/tcsh.  It leads to the
following error:

$ flux jobs -A -o "{contextual_time!H}"
H: Event not found.

While there is some general documentation on the need for escaping
and/or quoting, it doesn't seem to be explicit enough.

Solution: Add notes in flux-jobs(1) reminding some users that some
shells will specifically interpret braces and exclamation points, and
point out that they will need to be quoted and/or escaped.

Fixes #4938