job-manager: fix jobs stuck in PRIORITY when urgency is unchanged
Problem: The job manager prioritization code contains an optimization
that avoids posting a priority event if the priority did not change
after reprioritization. However, this causes jobs that are in PRIORITY
state to get stuck there, since they normally transition to SCHED
as a result of the priority event. This occurs if a job's urgency is
updated to the same value, e.g. if a held job is held again.

Force a priority event to be emitted in states other than SCHED, even
if the priority is unchanged. This allows jobs kicked back to PRIORITY
state to transition to SCHED, even if a duplicate urgency value is set.

Fixes #4940