libsubprocess: use a future for server shutdown
Problem: subprocess_server_subprocess_kill() re-enters the main
reactor and stays there for a timeout, thus it must be called late
in the host reactor's teardown to avoid letting other things run
that should be stopped.

Replace that function with subprocess_server_shutdown() which must
be called while the reactor is still running.  It returns a future
that is fulfilled when all subprocesses are cleaned up.

The broker registers a continuation on the future with a timeout.
After the future is fulfilled or times out, the broker reactor can
exit and destroy the subprocess server.