flux-mini: stash argparser parser in batch and alloc objects
Problem: The argparse objects for the flux-mini subcommand classes
are not available except in the main() function, which is inconvenient
if arguments need to be parsed again later.

For flux-mini 'batch' and 'alloc', set create the subparser as a
member of the cmd object so it can be accessed later.

This will allow the parser to be reused, if for example submission
directives are detected in the source file.