libflux: conf: allow path to file in flux_conf_parse()
Problem: flux_conf_parse() can only be used to parse a directory full
of files matching the *.toml glob pattern. This makes it inconvenient
when the configuration is in an individual file, since the file has
to be placed in a subdirectory with no other conf files.

If the path provided to flux_conf_parse() is not a directory, then
attempt to parse the path as an individual TOML file.