libsubprocess: drop EXEC_FAILED state
Problem: the FLUX_SUBPROCESS_EXEC_FAILED state is not needed.

This state is a terminal state that indicates that a remote
flux_exec() call failed before a PID could be returned.  Other
remote failures, before and after a PID is returned, are covered by
FLUX_SUBPROCESS_FAILED, so the extra state seems to serve no purpose.

Drop the state.

In addition, drop the function for decoding the rexec.exec
state=FAILED|EXEC_FAILED response that is no longer sent by the server.

Update tests and users.