doc: clarify documentation of options that set config PATH
Problem: There are a few places where `--conf-path` and `FLUX_CONF_DIR`
are documented in man pages and usage strings as taking only a
directory of TOML files, but this is no longer true. The underlying
flux_conf_parse() function now accepts a file as its argument, and
in this mode accepts either TOML or JSON.

Update places where `FLUX_CONF_DIR` and config-setting options are
documented to clarify that either of these modes of setting the conf
path may specify an individual file.