libsubprocess: improve errors for received subproc output
Problem: when the remote subprocess server returns output that
cannot be buffered, the error messages are not too helpful.

Promote a short write to the buffer to a fatal error and use ENOSPC
to match the error flux_buffer returns when completely full and
consolidate logging for short write and other buffer errors.

Provide more context in the error messages so it's clear which buffer
is overflowing.

Consolidate unlikely message decode errors into one log message.