python: drop [Error N] prefix for errors from CLIMain
Problem: Error messages from utilities using the CLIMain decorator
use the default string representation of OSError, which includes a
fairly useless `[Error N]` prefix.

If an exception handled by CLIMain() contains a "strerror" attribute,
log just that instead of the default strint representation of the
exception. Additionally handle file errors by appending ": {filename}"
if this attribute is found in the Excetption.

In the case of OSError and derived exceptions, this will avoid
printing the `[Error N]` prefix, but will otherwise keep the errors
mostly the same.