shell: batch: copy script to FLUX_JOB_TMPDIR
Problem: The shell batch plugin copies the batch script to the TMPDIR
of the job shell which could be a shared temporary filesystem. This
requires the use of mkstemp() to create a script file with a randomized
name. However, the shell already has a unique job temporary directory
created by the tmpdir plugin, and this may be a more appropriate
place for the batch script.

Place the batch script in `$FLUX_JOB_TMPDIR/script`. This removes
the need for mkstemp(3) and makes the location of the script more
predictable.