shell: batch: do not rewrite args that already run flux broker
Problem: The batch job shell plugin always rewrites the arguments of
a task to `flux broker /path/to/script` even if the command is already
in that form. By rewriting the task command line arbitrarily, the batch
plugin could overwrite arguments that were intended for the broker.

Check for an initial argument of "flux" and a second argument of
"broker" or "start". If this is the case then assume the job is
already a batch job and there is no need to rewrite arguments.