python: simplify Jobspec.from_batch_command()
Problem: The construction of "batch" jobspec is unnecessarily
different from jobspec created for interactive instance via
Jobspec.from_nest_command(). A batch jobspec is also a nested jobspec
which contains an embedded script to run instead of a command.

Now that the path to the script can be templated using {{tmpdir}}
in command arguments, a batch job can be constructed with arguments

 flux broker [broker_opts...] {{tmpdir}}/script [args...]

Rewrite Jobspec.from_batch_command() so that it is a special case of
from_nest_command(). Drop the `attributes.system.batch.broker-opts`
key since the broker options can now be passed directly to flux-broker.

Note that processing of this attribute along with rewriting args
in the job shell batch plugin cannot be removed for now in order to
continue processing jobs which may be in the queue after an upgrade.