testsuite: ensure shell batch plugin works with older jobspec
Problem: The construction of batch jobspec has changed, but there is
no test that ensures pending jobs from a previous version of flux-core
will work with the current job shell batch plugin.

Add a jobspec from flux-core v0.47.0 and a test to
t2613-job-shell-batch.t that ensures that and any other batch jobspecs
work in the current version of flux-core.