libsubprocess: fix SIGTERM on rexec error
Problem: although code exists to send the remote subprocess
a SIGKILL if the remote rexec.exec RPC terminates abnormally,
it is never executed because it is conditional on the state being
RUNNING right after the state is transitioned to FAILED.

Skip the RUNNING check and instead add a helper function that
calls remote_kill() only if the pid has been set.

Drop the logging on failure to send the rexec.kill RPC since that
is not terribly helpful.