libsubprocess: abstract logging with llog
Problem: libsubprocess code calls flux_log() internally,
but this is not appropriate in all situations.

Add flux_set_default_subprocess_log() which associates an llog
callback and llog argument with a flux_t handle.  When a subprocess
or the subprocess server is created, it sets its internal logger
to the one from the handle.  If this function is not called,
then the library does not confiugre a logger and will not log anything.

Convert all internal calls to flux_log() to llog counterparts.

Unconditionally compile libsubprocess with LLOG_ENABLE_DEBUG=1 so the
calls to llog_debug() are not compiled out.