python: OutputFormat: allow '?:' prefix to work with zero values
Problem: It may be useful to have output columns suppressed when all
values are "zero" (or equivalent), but the `?:` format prefix only applies
to empty strings.

Expand the set of "empty" values that are suppressed by '?:' in
OutputFormat.filter_empty().

e Please enter the commit message for your changes. Lines starting