cmd: add support for sections to flux(1) help output
Problem: The output of flux help is cluttered and thus may not be as
helpful as possible.

Split the output from flux help into sections. Support a new "tag"
in man page sources "flux-help-section" to allow the section of a
command to set at the same site in which it is added to help.

Rewrite the gen-cmdhelp.py script to organize included files into
sections, and include the section descriptions in the generated help
JSON config file.

Actual organization of manpages into flux help sections will be done
in a separate commit.

Adjust the flux-help extensibility test to use the new format.