configure: add --enable-broken-locale-mode to force ascii
Problem: Many sites appear to have locale and terminal
misconfigurations which cause the flux fancy ƒ character to
misrendered. This results in a large user impact since all jobids are
prefixed with this character. In these cases, the FLUX_F58_FORCE_ASCII
could be used, but it may be better to offer a definitive option.

Add a configure option to force the use of "f" instead of "ƒ" for the
project as a whole. This can then be enabled for sites that have known
locale or terminal configuration issues.