cmd: add -u, --unbuffered option to flux submission commands
Problem: There is no easy way to set the io buffering of jobs to
fully unbuffered, which may be a requirement for some jobs.

Add a new -u, --unbuffered option to flux submission commands, which
sets the output.stdout.buffer.type shell option to "none". This
option also shortens the output eventlog "batch timeout" to 50ms,
(from the default of 500ms) so that output is posted the eventlog
with only a very small delay.