flux-run: propagate --unbuffered option to flux job attach
Problem: When the --unbuffered option is used with flux-run, user
expectation is that standard input is also unbuffered, but this is
not currently the case.

Propagate the --unbuffered option to `flux job attach` when it is
utilized in `flux run`. This disables line buffering of stdin, so
that input will be sent to the job without waiting for a newline.