shell: input: do not disable input service on EOF
Problem: The shell shuts down its standard input service when it
receives a single EOF, so no more input can be sent. But the EOF
may have only been sent to a subset of tasks, in which case more
input may be expected.

Do not bother shutting down the shell input service.