testsuite: fix short hang in t2500-job-attach.t
Problem: The final test in t2500-job-attach.t hangs for about 5s
because the shell pty service is waiting for something to attach to
the interactive pty. (The job is finally killed when SIGKILL is sent
5s after SIGTERM is sent to the job).

Run `cat` instead of an interactive shell in the test, then after
the expected failure to attach with --read-only, terminate the job
by sending EOF with `runpty.py -i none ...` which will cause `cat` to
exit immediately, thus removing the need to cancel the job altogether.