shell: track tasks with open stdin
Problem: The job shell allows data sent to the input service to
be directed to task ranks whose stdin has already been closed.
This should generate an error.

Track tasks with stdin still open and compare this to the ranks in
the request. Respond with an error if the ranks is not a strict subset
of the open ranks (or is not the string "all")