testsuite: add tests for attach --stdin-ranks
Problem: No tests in the testsuite exercise behavior of the flux-job
attach --stdin-ranks option. Additionally, a corner case for handling
per-task input to the shell stdin service is not tested.

Add a test to ensure flux-job attach --stdin-ranks works.

Also ensure that sending stdin to a closed task fails as expected.