flux-run: pass --input=RANKS along to flux-job attach
Problem: Users have no way to direct stdin from flux-run to a subset
of tasks, but this is supported by flux-job attach.

If the --input option in flux-run specifies an idset, assume this is
a set of ranks to pass along to flux-job attach --stdin-ranks.

Since it could be confusing that an idset ends up trying to read from
a similarly named file with flux-submit, make it an error if an idset
is specified with the rest of the submission commands.