flux-job: improve handling of incorrect --stdin-ranks
Problem: The value for --stdin-ranks provided by the user is not
checked.  This can lead to errors at runtime which abort `flux job
attach` but leave the job running.

Check the value of --stdin-ranks at startup to ensure it is a valid
idset.

Also, ensure that the user-provided stdin ranks is a subset of all
task ranks in a job by comparing to the taskmap in the shell.start
event. If the user-provided idset has ranks that do not intersect
with the job task ranks, automatically adjust stdin_ranks and issue
a warning. This prevents an EPIPE error at runtime which would
cause `flux job attach` to exit while the job continues running.