libsubprocess: handle STOPPED state in the server
Problem: when a subprocess that was launched by the subprocess
server is stopped, the server throws an "illegal state" error
and destroys the subprocess.

Generate a state change response.
Call the user's on_state_change() callback.
Since STOPPED does not alter the subprocess state in the local
case, we don't alter it in the remote case either.

Fixes #4947