python: add copy method to OutputFormat
Problem: The ability to easily copy an instance of the OutputFormat
class, possibly with some fields removed, would be useful when
utilities wish to attempt deduplication of similar lines - but this
functionality does not exist.

Add a copy method to the OutputFormat class with an optional
`except_fields` parameter. This allows copying of format strings with
some fields removed.