job-manager: allow kill and killall for jobs in CLEANUP
Problem: If a job is in CLEANUP, the job-manager.kill and killall
RPCs will respond with EINVAL and the error "job is not running."
However, a job exception immediately transitions a job to CLEANUP,
which means the job shell and processes may still be running, so kill
and killall should be allowed for jobs in this state.

Change the test for job-manager.kill and job-manager.killall to

  if (!(job->state & FLUX_JOB_STATE_RUNNING))

to allow these RPCs to work for jobs that recently had an exception.