flux-mini: improve deprecation warning
Problem: The warning that flux-mini is deprecated is not as clear
as it could be.

Atttempt to instruct users the exact command they are looking for
when possible, e.g. `flux mini run` warns:

 flux-mini: WARNING: ⚠️ flux-mini is deprecated, use flux run ⚠️

Otherwise, replace 'flux-run`, 'flux-submit' with "flux run" and
"flux submit" so that users do not try to use literally "flux-run".

Fixes #4988