job-exec: avoid sending SIGKILL to job shell
Problem: When terminating a job due to a severity 0 exception, the
job exec system first sends SIGTERM to the job shell, which forwards
that signal on to the job tasks. If the job does not finish within
a configurable time limit (default=5s), then the exec system sends
SIGKILL directly to the job shell process group, terminating it
immediately. This can leave job processes running, for example if
tasks are ignoring SIGTERM.

Instead of sending SIGKILL to the job shell after the kill timer, send
a request to the job shell to send SIGKILL to the job. This is done
via `flux_job_kill(3)`. This should result in immediate termination
of the job tasks, which will cause the job shell to exit and the job
to complete.

There is still a potential for job tasks to be left around if tasks
escape the process group assigned by the job shell. At the system
instance level, this will be fixed in a future update to enable
containment.

Fixes #4987