python: Hostlist: return new Hostlist with index slice
Problem: When slicing a Hostlist object, a Python list is returned, but
it would be far more practical and efficient to return a new Hostlist
instead. Also, this follows the example of Python list slicing, since
indexing a list returns an element, but slicing returns a new list.

Update the Hostlist index method to return a new Hostlist when a
slice is used. Update the expand() method and one test accordingly.