testsuite: make jq(1) a global test requirement
Problem: The use of jq(1) is prevalent in the testsuite, but not
a requirement, so every use of jq(1) must be protected by HAVE_JQ
or a per-test global prereq check. The testsuite even creates some
jq wrappers to ensure that the jq command is not used without being
under the HAVE_JQ prereq. However, even with all that there are some
tests that fail without jq(1) because they use the products of jq(1)
from a previous test.

Given the prevalence of jq(1) usage in the testsuite, and the ease
with which mistakes in prerequisites may be made, just make jq(1)
a global requirement for the testsuite.

Remove all the testsuite helper code that attempted to make the use
of HAVE_JQ easy and less error prone.

Remove all references to HAVE_JQ and derivatives thereof.

Remove all uses of skip_all_unless_have(), which has been removed.