build: fix link dependencies for job-manager mods
Not sure why this only appears on nixos, but the submit-hold,
alloc-bypass and perilog modules were all missing necessary link
dependencies.  I think this was hiding behind our normal ignore missing
symbols behavior for modules and the load order on other systems
happening to load the libjob and libflux-internal libraries before the
modules on most platforms.