resource: drain: add force-overwrite mode
Problem: The resource module supports updating the drain reason of
a target via the drain "overwrite" mode, but there is no method to
also update the drain timestamp. This may be necessary if an admin
wishes to atomically redrain a node for a different reason, such that
the original timestamp no longer applies. If the admin has to first
undrain the node to update the timestamp, then there is a chance the
scheduler could assign the node to job while it is undrained.

Add a "force-overwrite" mode which causes the resource module to
overwrite both the drain reason and timestamp of targets.