flux-resource: drain: update drain timestamp with -ff
Problem: There is no way to set the "force-overwrite" mode in the
job-manager drain RPC from flux-resource drain.

If the drain --force option is used twice, update the mode in the
drain RPC from "overwrite" to "force-ovwrite", which requests that
the resource module update the drain timestamp along with any reason.

Fixes #4986