nix: add a flake.nix for flux build and dev.
The flake.nix file encodes all of flux core's dependencies, and enough
to build and install under NixOS. With the test changes, this passes all
but one test so far, and provides both a bare development environment
and a more full-featured "dev" shell environment with more tools.  In
principle this could be used to do build and caching of flux-core for
use elsewhere, but for now it's mainly here for discussion.