doc: reformat flux-resource(1) for clarity
Problem: The description of subcommands in the flux-resource(1)
man page is compact and therefore difficult to read.

Split descriptions of commands in flux-resource(1) into multiple lines
so it is easier to read. Give each option its own line for the most
part, and move more important options first.

Expand the description of the `status` command so it is clearer what
the command does.