broker: factor out common methods
Problem: the ping, rusage, and stats service method implementations
are not portable outside of the broker, but these may be useful in
other contexts such as the shell.

Move the message handlers to libfluxutil/method.[ch].
We'll create this local library as a catch-all for reusable bits
that depend on libflux-core.so.

Eliminate the broker specific code for each method that abstracted
message handler registration, and use flux_msg_handler_addvec_ex()
instead.

Since the ping method needs the server uuid and there is no general
way to obtain that, require "flux::uuid" to be set in the flux_t aux
container.