shell: add rexec service
Problem: tools may need to launch processes that are closely
coupled to a job's processes, but flux-exec(1) currently
launches processes that are direct children of flux-broker(1)
rather than flux-shell(1).  When more draconian job containment
is implemented in the future, this may be too decoupled.

As an example, STAT[1], which uses LaunchMON[2], has a Flux
plugin that uses flux-exec(1) to start the tool daemons.

Add a shell "rexec" plugin that embeds the libsubprocess server.
This adds the plumbing necessary for flux-exec(1) to launch
processes that are direct children of flux-shell(1).

This service is restricted to jobs running as the instance owner.
Furthermore, message credentials alone are not trusted to make this
determination.  The parent process is also verified to be running as
the same user.

[1] https://github.com/LLNL/STAT
[2] https://github.com/LLNL/LaunchMON

Co-authoried-by: Mark A. Grondona <mark.grondona@gmail.com>