testsuite: add job-exec SIGKILL/cleanup tests
Problem: No tests ensure that job cancellation via SIGKILL works
between the job-exec module and the shell when the kill timer expires.

Since there isn't a viable existing sharness test for this particular
issue, add a new test t2406-job-exec-cleanup.t. In the test, ensure
that a process which blocks SIGTERM is terminated by the job-exec
module's request to the shell to forward SIGKILL.