job-exec: terminate job shell with SIGKILL after timeout
Problem: The job exec module sends SIGKILL to force termination of a
job via the job-manager.kill service, which sends a shell-<jobid>.kill
event to request the job shells forward the signal to job tasks. This
depends upon a working job shell which is able to process this event.
If a job shell is stuck outside the reactor, or has been replaced with
a malicious shell which ignores this event, then the job may never post
a finish event (thus allowing a user extended access to resources).

Ensure the job shell is eventually killed by adding a second timer
which sends SIGKILL directly to the job shell after 5 times the kill
timeout (20s after the initial SIGKILL by default).