job-manager: fix post of free event during epilog
Problem: If a job exception is generated before a job starts, and
an epilog-start event is then posted in the CLEANUP state, a free
event will be posted before the epilog-finish event. This will cause
the job to become inactive before the epilog action finishes. This
occurs because the epilog-start event is enqueued in the job manager
event queue and thus the perilog_active refcount has not yet been
incremented.

In addition to the other checks in the CLEANUP state action, check
for a pending epilog-start event and do not emit the free event if
one is found.