testsuite: add check for epilog-start event after exception
Problem: No tests in the testsuite ensure that an epilog-start event
posted in the CLEANUP state for a job that never started holds the
free event until a corresponding epilog-finish event.

Enhance the perilog-test.so plugin to optionally generate a job
exception while the prolog is active. Use this test to ensure that
the epilog-start event prevents the job manager from posting the
free event.