testsuite: fix python failure in t2613-job-shell-batch.t
Problem: The final test in t2613-job-shell-batch.t attempts to run a
jobspec generated by `flux batch` from an older version of flux-core.
While the test does update PATH in the environment of the older
jobspec, it fails to update PYTHONPATH and so can fail depending on
the environment in which the test is being run.

Update PYTHONPATH as well as PATH in this test.