testsuite: fix spurious failures in flux-help tests
Problem: Under a nix pkg environment, flux-help tests in t0001-basic.t
fail due to errors like:

  Bad system call (core dumped)
  man: col: Bad system call (core dumped)

This appears to be due to seccomp sandboxing in man-db, and
experimentally exporting MAN_DISABLE_SECCOMP=1 works around the issue.

Since security is less of a concern during this testsuite, just export
MAN_DISABLE_SECCOMP=1 for the duration of the test.