testsuite: fix spurious failure in t2405-job-exec-sdexec.t
Problem: The test 'job-exec: simple job outputs stderr under systemd'
fails if any other output is sent to stderr during the test. This
should not cause a failure.

Instead of testing exact output, just grep for the expected output
in this test. This effectively ignores any extra logging or warnings
to stderr.