cli: adjust description of --begin-time submission option
Problem: The help and documentation of the submission --begin-time
option does not make it clear enough that an offset can be provided
to the option via a '+' character.

Adjust the help string and man page documentation to clarify that the
option takes either an offset in FSD or an absolute datetime.