flux-top: fix display issue with queue filter
Problem: When the queue filter is used (i.e. --queue=name), some jobs
can be displayed on the wrong line, because the index for all jobs
(i.e. without queue filtering) is used to determine the line for display.

Solution: Use a counter to determine the relative line number that
each job should be displayed on.

Fixes #5021