resource: drain: do not allow drain of excluded ranks
Problem: `flux resource status` assumes ranks are only in one of the
states exclude, draining, or drained, but currently ranks can be both
excluded and drained. Since the purpose of draining resources is to
remove them from consideration by the schedueler, it isn't necessary
to support draining of excluded nodes.

Do not allow currently excluded ranks to be drained in the resource
module.