resource: drain: ignore drain state of excluded ranks on replay
Problem: A rank which was drained before a restart could be
subsequently excluded by configuration or other means. This would
allow the rank to be both excluded and drained, but this is assumed
to never be true by tools such as `flux resource status`.

Ignore excluded ranks in the drain replay code to avoid draining them.