resource: allow an empty resource.exclude idset
Problem: If an empty idset is specified in the resource.exclude
configuration, the resource module complains that the exclude idset is
"out of range". This occurs because the resource module is checking
the range with idset_last(3), which returns IDSET_INVALID_ID for an
empty idset.

Check if the exclude idset is empty before testing the range with
idset_last(3).