testsuite: expand resource module drain tests
Problem: The resource module drain tests do not include any tests of
attempting to drain an excluded rank, or handling of a rank which has
been excluded by configuration but was drained before a restart, etc.

Expand the drain testing in t2311-resource-drain.t to test some corner
cases with exclude and drain.