flux-pgrep: correct default override corner cases
Problem: flux-pgrep has several unique output variations compared to
other tools that make the FLUX_PGREP_FORMAT_DEFAULT environment
variable not behave as expected.

Solution: Set the tool name "flux-pgrep" when initializing the configuration
class.  This is so the correct environment variable FLUX_PGREP_FORMAT_DEFAULT is
searched for.  Ensure header is output if the overridden default output
is not the same as the builtin default.