flux-ping: add header before ping output
Problem: The ping(1) command outputs a header before the ping responses.
The header can be be useful to inform the user what they are pinging, like
the IP address if a hostname was input on the command line.  A similar
header can be useful for flux-ping when a hostname is input (notify user
of the rank) or if the user does not specify a rank, that the rank "any"
is used.

Solution: Add a header output before the ping responses.

Fixes #4106