flux-ping: make output consistent
Problem: The output from flux-ping is different depending on
the input on the command line, most notably when the input
from the user is "any" or "upstream".

Solution: Have the ping handler return its rank in the ping response.
Use that rank in all ping output to make the output consistent
regardless of the input.