resource: drain: add function to undrain a set of ranks
Problem: An internal function to drain ranks is exposed to other
resource module subsystems in drain.h, but no such function is exposed
to undrain ranks. This will be necessary when nodes are dynamically
excluded.

Add an undrain_ranks() function to drain.h for use by resource module
users outside of drain.c.