job-manager: support job-manager.purge-id target
Problem: The current job-manager.purge target purges a range of job ids
based on age or a numerical limit.  However, there may be specific
circumstances that the user may want to purge a single job.  For example,
if a single job is corrupted and has invalid data.

Solution: Add a new job-manager.purge-id target that takes a single job id
as input for purging.