flux-job: support passing job ids to purge
Problem: Sometimes it would be nice to purge a specific job id, instead
of a range of job ids.  This is typically necessary in error or recovery
situations.

Support passing job ids to the purge subcommand.  When IDs are passed in,
they are sent to the job-manager.purge-id target instead job-manager.purge.

Fixes #4801