python: add a to_dict() method to the JobInfo class
Problem: There's no easy way to convert a JobInfo object representing
a job to a dictionary. This would be useful since JobInfo has a set
of derived attributes which are more interesting than the raw data
returned directly by job-list.

Add a to_dict() method to the JobInfo class. The set of fields added
to this dict are defined by the keys of the `defaults` class attribute
along with a new `properties` attribute, plus a few attributes that
require special handling for conversion into a dict.

By default, filter unset and invalid attributes from the returned
dict to avoid confusion. An unfiltered dictionary can be obtained by
passing filtered=False to the to_dict() method.