libpmi: add upmi plugin for libpmi2.so
Problem: At LLNL, the default environment sets LD_LIBRARY_PATH
to include the directory containing Cray's PMI libraries, but
the Cray libpmi.so cannot be used to bootstrap Flux when Slurm
launches Flux or when a singleton is launched with flux start.

There are two problems with cray's libpmi.so:
- it hangs when asked for keys that do not exist
- it truncates our public keys during the exchange

The Cray libpmi2.so works better, so add a upmi plugin for libpmi2
and place it in the default method list before libpmi.

Fixes #5040