flux-job: add special exit code to flux job wait
Problem: A common use case is to use flux job wait in a loop
and wait until all waitable jobs are complete.  Flux job wait will
have an exit code of 1 when there are no more waitable jobs.  It
also has an exit code of 1 when a waitable job exits with an error.
This inconsistency makes it difficult to script against flux job wait.

Solution: When flux job wait is used without the --all option, exit
with the exit code of 2 if the job id is invalid or the jobid is not
waitable or no more waitable jobs exist.

Fixes #5033